Fixed installation file
Requirements.txt had some old libraries and conflicting packages. This has been resolved for the most part. I still need to develop a python script that will handle some of the manual steps that are still there. But this should be a huge help in resolving installation problems even as is.