Setup development environment
I've enabled the code to be run in development mode as well as included a similar script to be able to quickly start the application in production as well. It tries to update the dependencies as well but this needs to be looked into. I need to iron out some of the problems of package interdepencies as well.