Fixed status message display
Because some situations had a rapid number of status updates being sent from the backend to the frontend, only the latest message was shown. Now each update is shown on the frontend for 250ms. A status message queue was implemented.

Also updated the start script to launch and be accessible on a LAN.