MCP server implemented
Currently only two tools are present -- math_tool and get_conversation_history. The remaining six tools still need to be implemented. But recent context is maintained which makes having a conversation with this agent nicer.

The frontend is also updated to include a Thinker agent now that uses these MCP tools specifically. Very good formatting is now available on the frontend.

Favicon has now been added. But it needs to be made better.