Got rid of dependency conflicts
Remove atproto and farcaster as they have not been upgraded to work with the latest websockets versions which conflicts with FastMCP 2.0. We will need to figure out different ways to use these applications.

Now also included a setup script to make it easy to install the app. Mac and Windows versions are both there.