Updated the setup scripts
It now installs each of the packages one at a time and if it runs into dependency conflicts or issues with any one package, it carries on to the next one. This will allow the application to at least begin even though it may cause issues when conflicting packages render some minor feature unusable.